+++ smarter/smarter/apps/api/v1/cli/views/base.py
@@ -5,10 +5,8 @@ import logging
 import re
 from http import HTTPStatus
 from typing import Type
-from urllib.parse import ParseResult, urlparse
 
 import yaml
-from django.core.handlers.wsgi import WSGIRequest
 from django.http import QueryDict
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.views import APIView
@@ -17,8 +15,8 @@ from smarter.apps.account.mixins import AccountMixin
 from smarter.apps.api.v1.cli.brokers import Brokers
 from smarter.apps.api.v1.manifests.enum import SAMKinds
 from smarter.apps.api.v1.manifests.version import SMARTER_API_VERSION
-from smarter.common.classes import SmarterHelperMixin
 from smarter.common.exceptions import SmarterExceptionBase
+from smarter.lib.django.request import SmarterRequestMixin
 from smarter.lib.django.validators import SmarterValidator
 from smarter.lib.drf.token_authentication import SmarterTokenAuthentication
 from smarter.lib.journal.enum import SmarterJournalCliCommands
@@ -47,7 +45,7 @@ class APIV1CLIViewError(SmarterExceptionBase):
 
 
 # pylint: disable=too-many-instance-attributes
-class CliBaseApiView(APIView, AccountMixin, SmarterHelperMixin):
+class CliBaseApiView(APIView, SmarterRequestMixin):
     """
     Smarter API command-line interface Base class API view. Handles
     common tasks for all /api/v1/cli views:
@@ -60,7 +58,6 @@ class CliBaseApiView(APIView, AccountMixin, SmarterHelperMixin):
 
     authentication_classes = (SmarterTokenAuthentication,)
     permission_classes = (IsAuthenticated,)
-    request: WSGIRequest = None
 
     _BrokerClass: Type[AbstractBroker] = None
     _broker: AbstractBroker = None
@@ -71,6 +68,10 @@ class CliBaseApiView(APIView, AccountMixin, SmarterHelperMixin):
     _manifest_load_failed: bool = False
     _params: dict[str, any] = None
     _prompt: str = None
+
+    def __init__(self, *args, **kwargs):
+        APIView.__init__(self, *args, **kwargs)
+        SmarterRequestMixin.__init__(self, *args, **kwargs)
 
     @property
     def loader(self) -> SAMLoader:
@@ -105,7 +106,7 @@ class CliBaseApiView(APIView, AccountMixin, SmarterHelperMixin):
         parameters to the broker.
         """
         if not self._params:
-            self._params = QueryDict(self.request.META.get("QUERY_STRING", "")) or {}
+            self._params = QueryDict(self.smarter_request.META.get("QUERY_STRING", "")) or {}
         return self._params
 
     @property
@@ -132,7 +133,7 @@ class CliBaseApiView(APIView, AccountMixin, SmarterHelperMixin):
         if self.BrokerClass and not self._broker:
             BrokerClass = self.BrokerClass
             self._broker = BrokerClass(
-                request=self.request,
+                request=self.smarter_request,
                 api_version=SMARTER_API_VERSION,
                 name=self.manifest_name,
                 kind=self.manifest_kind,
@@ -210,27 +211,15 @@ class CliBaseApiView(APIView, AccountMixin, SmarterHelperMixin):
         instance. For example, if the route is '/api/v1/cli/apply/', then
         the corresponding command will be SmarterJournalCliCommands.APPLY.
 
-        url: http://testserver/api/v1/cli/logs/Chatbot/?name=TestChatBot
+        url:
+         - http://testserver/api/v1/cli/logs/Chatbot/?name=TestChatBot
+         - http://testserver/api/v1/cli/apply
         """
         match = re.search(r"/cli/([^/]+)/", self.url or "")
         if match:
             _command = match.group(1)
             return SmarterJournalCliCommands(_command)
         raise APIV1CLIViewError(f"Could not determine command from url: {self.url}")
-
-    @property
-    def url(self) -> str:
-        """
-        Get the full url of the request. Reconstructs the exact url of
-        the request. example url:
-
-        https://platform.smarter.sh/api/v1/cli/chat/config/example/?new_session=false&uid=Lawrences-Mac-Studio.local-c6%253A6b%253A2e%253A7a%253A3d%253A6c
-        """
-        return SmarterValidator.urlify(self.request.build_absolute_uri())
-
-    @property
-    def url_parsed(self) -> ParseResult:
-        return urlparse(self.url) if self.url else None
 
     # pylint: disable=too-many-return-statements,too-many-branches
     def dispatch(self, request, *args, **kwargs):
@@ -263,13 +252,17 @@ class CliBaseApiView(APIView, AccountMixin, SmarterHelperMixin):
             It provides a service interface that 'brokers' the http request for the
             underlying object that provides the object-specific service (create, update, get, delete, etc).
         """
-        self.request: WSGIRequest = request
+
+        # this is a hacky way to get SmarterRequestMixin request object
+        # initialized.
+        super().init(request=request)
+
         # Manifest parsing and broker instantiation are lazy implementations.
         # So for now, we'll only set the private class variable _manifest_data
         # from the request body, and then we'll leave it to the child views to
         # decide if/when to actually parse the manifest and instantiate the broker.
         try:
-            data = self.request.body.decode("utf-8")
+            data = request.body.decode("utf-8")
             # if the command is 'chat', then the raw prompt text
             # or the encoded file attachment data will be in the request body.
             # otherwise, the request body should contain manifest text.
@@ -294,12 +287,12 @@ class CliBaseApiView(APIView, AccountMixin, SmarterHelperMixin):
                     )
 
         try:
-            AccountMixin.__init__(self, user=self.request.user)
+            AccountMixin.__init__(self, user=request.user)
             if self.user.is_authenticated and not self.user_profile:
                 raise APIV1CLIViewError("Could not find account for user.")
         except SmarterExceptionBase as e:
             return SmarterJournaledJsonErrorResponse(
-                request=self.request,
+                request=request,
                 thing=self.manifest_kind,
                 command=self.command,
                 e=e,
@@ -310,19 +303,19 @@ class CliBaseApiView(APIView, AccountMixin, SmarterHelperMixin):
         self._manifest_name = self.params.get("name", None)
 
         if self.authentication_classes and self.permission_classes:
-            if not self.request.user.is_authenticated:
+            if not request.user.is_authenticated:
                 try:
                     raise APIV1CLIViewError("Unauthorized access attempted.")
                 except APIV1CLIViewError as e:
                     return SmarterJournaledJsonErrorResponse(
-                        request=self.request,
+                        request=request,
                         thing=self.manifest_kind,
                         command=self.command,
                         e=e,
                         status=HTTPStatus.FORBIDDEN.value,
                     )
 
-        user_agent = self.request.headers.get("User-Agent", "")
+        user_agent = request.headers.get("User-Agent", "")
         if "Go-http-client" not in user_agent:
             logger.warning("The User-Agent is not a Go lang application: %s", user_agent)
 
@@ -348,10 +341,10 @@ class CliBaseApiView(APIView, AccountMixin, SmarterHelperMixin):
         # to the super class dispatch method, and the parameters are passed
         # to the child view's post method.
         try:
-            return super().dispatch(self.request, *args, **{**self.params, **kwargs})
+            return super().dispatch(request, *args, **{**self.params, **kwargs})
         except SAMBrokerErrorNotImplemented as not_implemented_error:
             return SmarterJournaledJsonErrorResponse(
-                request=self.request,
+                request=request,
                 thing=self.manifest_kind,
                 command=self.command,
                 e=not_implemented_error.get_formatted_err_message,
@@ -359,7 +352,7 @@ class CliBaseApiView(APIView, AccountMixin, SmarterHelperMixin):
             )
         except SAMBrokerErrorNotReady as not_ready_error:
             return SmarterJournaledJsonErrorResponse(
-                request=self.request,
+                request=request,
                 thing=self.manifest_kind,
                 command=self.command,
                 e=not_ready_error.get_formatted_err_message,
@@ -367,7 +360,7 @@ class CliBaseApiView(APIView, AccountMixin, SmarterHelperMixin):
             )
         except SAMBrokerErrorNotFound as not_found_error:
             return SmarterJournaledJsonErrorResponse(
-                request=self.request,
+                request=request,
                 thing=self.manifest_kind,
                 command=self.command,
                 e=not_found_error.get_formatted_err_message,
@@ -375,7 +368,7 @@ class CliBaseApiView(APIView, AccountMixin, SmarterHelperMixin):
             )
         except SAMBrokerReadOnlyError as read_only_error:
             return SmarterJournaledJsonErrorResponse(
-                request=self.request,
+                request=request,
                 thing=self.manifest_kind,
                 command=self.command,
                 e=read_only_error.get_formatted_err_message,
@@ -383,7 +376,7 @@ class CliBaseApiView(APIView, AccountMixin, SmarterHelperMixin):
             )
         except SAMBrokerError as broker_error:
             return SmarterJournaledJsonErrorResponse(
-                request=self.request,
+                request=request,
                 thing=self.manifest_kind,
                 command=self.command,
                 e=broker_error.get_formatted_err_message,
@@ -392,7 +385,7 @@ class CliBaseApiView(APIView, AccountMixin, SmarterHelperMixin):
         # pylint: disable=broad-except
         except Exception as e:
             return SmarterJournaledJsonErrorResponse(
-                request=self.request,
+                request=request,
                 thing=self.manifest_kind,
                 command=self.command,
                 e=e,